refactor countLinesMatchingRegex into FileLineMatcher struct
